combine tests in a single file
